requirements.txt
fastapi
uvicorn
python-dotenv
linkedin-api
pydantic
aioredis
requests
aiohttp
tenacity